Move FailsWithMessage and FailsWithMessageExtension
