+++ subprojects/internal-testing/src/main/groovy/org/gradle/util/internal/FailsWithMessageExtension.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.util;
+package org.gradle.util.internal;
 
 import junit.framework.ComparisonFailure;
 import org.spockframework.runtime.ConditionFailedWithExceptionError;
